Sample log posterior and gradient functions for select generalized linear models and mixed effect models

These functions can be used to fit common generalized linear models and mixed effect models. See the accompanying vignettes for details on the derivations of the log posterior and gradient. In addition, these functions can be used as templates to build custom models to fit using HMC.

# extract parameters from theta vector beta_param <- as.numeric(theta[1:p]) tau_param <- theta[(p+1):(p+m*q)]

# extract parameters from theta vector beta_param <- as.numeric(theta[1:p]) tau_param <- theta[(p+1):(p+m*q)]

# prior covariance for beta Sig_inv_beta <- diag(1/sig2beta, p, p)

# prior covariance for beta Sig_inv_beta <- diag(1/sig2beta, p, p)

linear_posterior(theta, y, X, a = 1e-04, b = 1e-04, sig2beta = 1000)

g_linear_posterior(theta, y, X, a = 1e-04, b = 1e-04, sig2beta = 1000)

logistic_posterior(theta, y, X, sig2beta = 100)

g_logistic_posterior(theta, y, X, sig2beta = 100)

poisson_posterior(theta, y, X, sig2beta = 100)

g_poisson_posterior(theta, y, X, sig2beta = 100)

lmm_posterior(
  theta,
  y,
  X,
  Z,
  n,
  d,
  nrandom = 1,
  nugamma = 1,
  nuxi = 1,
  Agamma = 25,
  Axi = 25,
  sig2beta = 1000
)

g_lmm_posterior(
  theta,
  y,
  X,
  Z,
  n,
  d,
  nrandom = 1,
  nugamma = 1,
  nuxi = 1,
  Agamma = 25,
  Axi = 25,
  sig2beta = 1000
)

glmm_bin_posterior(
  theta,
  y,
  X,
  Z,
  n,
  nrandom = 1,
  nuxi = 1,
  Axi = 25,
  sig2beta = 10000
)

g_glmm_bin_posterior(
  theta,
  y,
  X,
  Z,
  n,
  nrandom = 1,
  nuxi = 1,
  Axi = 25,
  sig2beta = 10000
)

glmm_poisson_posterior(
  theta,
  y,
  X,
  Z,
  m,
  q = 1,
  A = 10000,
  nulambda = 1,
  Alambda = 25,
  sig2beta = 10000
)

g_glmm_poisson_posterior(
  theta,
  y,
  X,
  Z,
  m,
  q = 1,
  A = 10000,
  nulambda = 1,
  Alambda = 25,
  sig2beta = 10000
)

Arguments

theta
vector of parameters. Stored as a single vector in order fixed effect, random effect, log-transformed diagonal \(\lambda\), and off-diagonal of G vector a

y
numeric vector for the dependent variable

X
numeric design matrix of fixed effect parameters

a
hyperprior for the Inverse Gamma shape parameter

b
hyperprior for the Inverse Gamma scale parameter

sig2beta
diagonal covariance of prior for linear predictors is multivariate normal with mean 0

Z
numeric design matrix of random effect parameters

n
number of subjects for mixed effect models, or number of observations for standard glm

d
number of observations per subject

nugamma
hyperprior \(\nu\) for the half-t prior of the log transformed error

nuxi
hyperprior \(\nu\) for the half-t prior of the random effects diagonal

Agamma
hyperprior \(A\) for the half-t prior of the log transformed error

Axi
hyperprior \(A\) for the half-t prior of the random effects diagonal

m
number of random effect linear parameters

q
number of random effects covariance parameters

A
hyperprior numeric vector for the random effects off-diagonal a

nulambda
hyperprior \(\nu\) for the half-t prior of the random effects diagonal

Alambda
hyperprior \(A\) for the half-t prior of the random effects diagonal

nueps
hyperprior \(\nu\) for the half-t prior of the error parameter

Aeps
hyperprior \(A\) for the half-t prior of the error parameter

Value

Numeric vector for the log posterior or gradient of the log posterior

Details

Generalized Linear Models with available posterior and gradient functions

`linear_posterior(theta, y, X, a=1e-4, b=1e-4, B=0.001)`

The log posterior function for linear regression $$p(y | X, \beta; \sigma^2) = \frac{1}{(2\pi\sigma^2)^{n/2}}\exp{\left(-\frac{1}{2\sigma^2} (y - X\beta)^T(y-X\beta) \right)}$$ with priors \(p(\sigma^2) \sim IG(a, b)\) and \(\beta \sim N(0, BI)\). The variance term is log transformed \(\gamma = \log\sigma^2\) The input parameter vector theta is of length k. The first k-1 parameters are for \(\beta\), and the last parameter is \(\gamma\) Note that the Inverse Gamma prior can be problematic for certain applications with low variance. See Gelman (2006)

`g_linear_posterior(theta, y, X, a = 1e-04, b = 1e-04, B = 0.001)`

Gradient of the log posterior for a linear regression model with Normal prior for the linear parameters and Inverse Gamma for the error term. $$p(y | X, \beta; \sigma^2) = \frac{1}{(2\pi\sigma^2)^{n/2}}\exp{\left(-\frac{1}{2\sigma^2} (y - X\beta)^T(y-X\beta) \right)}$$ with priors \(p(\sigma^2) \sim IG(a, b)\) and \(\beta \sim N(0, BI)\). The variance term is log transformed \(\gamma = \log\sigma^2\) The input parameter vector theta is of length k. The first k-1 parameters are for \(\beta\), and the last parameter is \(\gamma\) Note that the Inverse Gamma prior can be problematic for certain applications with low variance. See Gelman (2006)

`logistic_posterior(theta, y, X, B = 0.01) `

Log posterior for a logistic regression model with Normal prior for the linear parameters. The likelihood function for logistic regression $$L(\beta; X, y) = \prod_{i=1}^{n} \left(\frac{1}{1+e^{-X_i\beta}}\right)^{y_i} \left(\frac{e^{-X_i\beta}}{1+e^{-X_i\beta}}\right)^{1-y_i}$$ with priors \(\beta \sim N(0, BI)\). The input parameter vector theta is of length k, containing parameter values for \(\beta\)

`g_logistic_posterior(theta, y, X, B = 0.01) `

Gradient of the log posterior for a logistic regression model with Normal prior for the linear parameters. The likelihood function for logistic regression $$L(\beta; X, y) = \prod_{i=1}^{n} \left(\frac{1}{1+e^{-X_i\beta}}\right)^{y_i} \left(\frac{e^{-X_i\beta}}{1+e^{-X_i\beta}}\right)^{1-y_i}$$ with priors \(\beta \sim N(0, BI)\). The input parameter vector theta is of length k, containing parameter values for \(\beta\)

`poisson_posterior(theta, y, X, B = 0.01) `

Log posterior for a Poisson regression model with Normal prior for the linear parameters. The likelihood function for poisson regression $$L(\beta; y, X) = \prod_{i=1}^n \frac{e^{-e^{X_i\beta}}e^{y_iX_i\beta}}{y_i!}$$ with priors \(\beta \sim N(0, BI)\). The input parameter vector theta is of length k, containing parameter values for \(\beta\)

`g_poisson_posterior(theta, y, X, B = 0.01) `

Gradient of the log posterior for a Poisson regression model with Normal prior for the linear parameters. The likelihood function for poisson regression $$L(\beta; y, X) = \prod_{i=1}^n \frac{e^{-e^{X_i\beta}}e^{y_iX_i\beta}}{y_i!}$$ with priors \(\beta \sim N(0, BI)\). The input parameter vector theta is of length k, containing parameter values for \(\beta\)

Generalized Linear Mixed Effect with available posterior and gradient functions

`lmm_posterior(theta, y, X, Z, m, q = 1, A = 10000, nueps = 1, nulambda = 1, Aeps = 25, Alambda = 25, B = 0.001) `

The log posterior function for linear mixed effects regression $$p(y | \beta, u, \sigma_\epsilon^2) \propto (\sigma_\epsilon^2)^{-n/2} e^{-\frac{1}{2\sigma_\epsilon^2}(y - X\beta - Zu)^T (y - X\beta - Zu)}$$ with priors \(\beta \sim N(0, BI)\), \(\sigma_\epsilon \sim half-t(A_\epsilon, nu_\epsilon)\), \(\lambda \sim half-t(A_\lambda, nu_\lambda )\). The vector \(\lambda\) is the diagonal of the covariance G hyperprior where \(u \sim N(0, G\). The off-diagonal hyperpriors are stored in a vector \(a \sim N(0, A\). See Chan, Jeliazkov (2009) for details. The input parameter vector theta is of length k. The first k-1 parameters are for \(\beta\), and the last parameter is \(\gamma\)

`g_lmm_posterior(theta, y, X, Z, m, q = 1, A = 10000, nueps = 1, nulambda = 1, Aeps = 25, Alambda = 25, B = 0.001)`

Gradient of the log posterior for a linear mixed effects regression model $$p(y | \beta, u, \sigma_\epsilon^2) \propto (\sigma_\epsilon^2)^{-n/2} e^{-\frac{1}{2\sigma_\epsilon^2}(y - X\beta - Zu)^T (y - X\beta - Zu)}$$ with priors \(\beta \sim N(0, BI)\), \(\sigma_\epsilon \sim half-t(A_\epsilon, nu_\epsilon)\), \(\lambda \sim half-t(A_\lambda, nu_\lambda )\). The vector \(\lambda\) is the diagonal of the covariance G hyperprior where \(u \sim N(0, G\). The off-diagonal hyperpriors are stored in a vector \(a \sim N(0, A\). See Chan, Jeliazkov (2009) for details. The input parameter vector theta is of length k. The first k-1 parameters are for \(\beta\), and the last parameter is \(\gamma\)

`glmm_bin_posterior(theta, y, X, Z, m, q = 1, A = 10000, nulambda = 1, Alambda = 25, B = 10000)`

The log posterior function for logistic mixed effects regression $$p(y | X, Z, \beta, u) = \prod_{i=1}^n\prod_{j=1}^m \left(\frac{1}{1 + e^{-X_{i}\beta - Z_{ij}u_i}}\right)^{y_{ij}} \left(\frac{e^{-X_i\beta - Z_{ij}u_i}}{1 + e^{-X_{i}\beta - Z_{ij}u_i}}\right)^{1-y_{ij}} $$ with priors \(\beta \sim N(0, BI)\), \(\sigma_\epsilon \sim half-t(A_\epsilon, nu_\epsilon)\), \(\lambda \sim half-t(A_\lambda, nu_\lambda )\). The vector \(\lambda\) is the diagonal of the covariance G hyperprior where \(u \sim N(0, G\). The off-diagonal hyperpriors are stored in a vector \(a \sim N(0, A\). See Chan, Jeliazkov (2009) for details. The input parameter vector theta is of length k. The first k-1 parameters are for \(\beta\), and the last parameter is \(\gamma\)

`g_glmm_bin_posterior(theta, y, X, Z, m, q = 1, A = 10000, nulambda = 1, Alambda = 25, B = 10000) `

Gradient of the log posterior function for logistic mixed effects regression $$p(y | X, Z, \beta, u) = \prod_{i=1}^n\prod_{j=1}^m \left(\frac{1}{1 + e^{-X_{i}\beta - Z_{ij}u_i}}\right)^{y_{ij}} \left(\frac{e^{-X_i\beta - Z_{ij}u_i}}{1 + e^{-X_{i}\beta - Z_{ij}u_i}}\right)^{1-y_{ij}} $$ with priors \(\beta \sim N(0, BI)\), \(\sigma_\epsilon \sim half-t(A_\epsilon, nu_\epsilon)\), \(\lambda \sim half-t(A_\lambda, nu_\lambda )\). The vector \(\lambda\) is the diagonal of the covariance G hyperprior where \(u \sim N(0, G\). The off-diagonal hyperpriors are stored in a vector \(a \sim N(0, A\). See Chan, Jeliazkov (2009) for details. The input parameter vector theta is of length k. The first k-1 parameters are for \(\beta\), and the last parameter is \(\gamma\)

`glmm_poisson_posterior(theta, y, X, Z, m, q = 1, A = 10000, nulambda = 1, Alambda = 25, B = 10000) `

Log posterior for a Poisson mixed effect regression $$L(\beta; y, X) = \prod_{i=1}^n \prod_{j=1}^m \frac{e^{-e^{X_i\beta + Z_{ij}u_{ij}}}e^{y_i(X_i\beta + Z_{ij}u_{ij})}}{y_i!} $$ with priors \(\beta \sim N(0, BI)\), \(\sigma_\epsilon \sim half-t(A_\epsilon, nu_\epsilon)\), \(\lambda \sim half-t(A_\lambda, nu_\lambda )\). The vector \(\lambda\) is the diagonal of the covariance G hyperprior where \(u \sim N(0, G\). The off-diagonal hyperpriors are stored in a vector \(a \sim N(0, A\). See Chan, Jeliazkov (2009) for details. The input parameter vector theta is of length k. The first k-1 parameters are for \(\beta\), and the last parameter is \(\gamma\)

`g_glmm_poisson_posterior(theta, y, X, Z, m, q = 1, A = 10000, nulambda = 1, Alambda = 25, B = 10000) `

Gradient of the log posterior for a Poisson mixed effect regression $$L(\beta; y, X) = \prod_{i=1}^n \prod_{j=1}^m \frac{e^{-e^{X_i\beta + Z_{ij}u_{ij}}}e^{y_i(X_i\beta + Z_{ij}u_{ij})}}{y_i!} $$ with priors \(\beta \sim N(0, BI)\), \(\sigma_\epsilon \sim half-t(A_\epsilon, nu_\epsilon)\), \(\lambda \sim half-t(A_\lambda, nu_\lambda )\). The vector \(\lambda\) is the diagonal of the covariance G hyperprior where \(u \sim N(0, G\). The off-diagonal hyperpriors are stored in a vector \(a \sim N(0, A\). See Chan, Jeliazkov (2009) for details. The input parameter vector theta is of length k. The first k-1 parameters are for \(\beta\), and the last parameter is \(\gamma\)

References

Gelman, A. (2006). Prior distributions for variance parameters in hierarchical models (comment on article by Browne and Draper). Bayesian analysis, 1(3), 515-534.

Chan, J. C. C., & Jeliazkov, I. (2009). MCMC estimation of restricted covariance matrices. Journal of Computational and Graphical Statistics, 18(2), 457-480.

Betancourt, M., & Girolami, M. (2015). Hamiltonian Monte Carlo for hierarchical models. Current trends in Bayesian methodology with applications, 79, 30.

Examples

# Linear regression example
set.seed(521)
X <- cbind(1, matrix(rnorm(300), ncol=3))
betavals <- c(0.5, -1, 2, -3)
y <- X%*%betavals + rnorm(100, sd=.2)

f1_hmc <- hmc(N = 500,
          theta.init = c(rep(0, 4), 1),
          epsilon = 0.01,
          L = 10,
          logPOSTERIOR = linear_posterior,
          glogPOSTERIOR = g_linear_posterior,
          varnames = c(paste0("beta", 0:3), "log_sigma_sq"),
          param=list(y=y, X=X), parallel=FALSE, chains=1)

summary(f1_hmc, burnin=100)

#> Summary of MCMC simulation
#> 

#>                      5%       25%        50%        75%        95%
#> beta0         0.4847789  0.518149  0.5321988  0.5474873  0.5726653
#> beta1        -1.0398451 -1.022090 -1.0099795 -0.9969146 -0.9695882
#> beta2         1.9834760  2.003620  2.0201366  2.0330953  2.0521270
#> beta3        -3.0146309 -2.996771 -2.9813806 -2.9668917 -2.9442674
#> log_sigma_sq -3.2522864 -3.150465 -3.0570739 -2.9441514 -2.8255475

# poisson regression example
set.seed(7363)
X <- cbind(1, matrix(rnorm(40), ncol=2))
betavals <- c(0.8, -0.5, 1.1)
lmu <- X %*% betavals
y <- sapply(exp(lmu), FUN = rpois, n=1)

f2_hmc <- hmc(N = 500,
          theta.init = rep(0, 3),
          epsilon = 0.01,
          L = 10,
          logPOSTERIOR = poisson_posterior,
          glogPOSTERIOR = g_poisson_posterior,
          varnames = paste0("beta", 0:2),
          param = list(y=y, X=X),
          parallel=FALSE, chains=1)